farm_rental_system/requirements.txt
pip~=24.2
DateTime~=5.4
wheel~=0.43.0
utils~=1.0.2
pillow~=10.2.0
Jinja2~=3.1.3
filelock~=3.13.1
redis~=5.0.3
Django~=4.2
djangorestframework~=3.14.0
django-author~=1.2.0
publisher~=0.11
django-publisher~=0.0.3
requests~=2.31.0
django-email-users~=1.1.0
mysql~=0.0.3
mysql-connector-python~=8.0.26
djangorestframework-jwt~=1.11.0
PyJWT~=1.7.1
typing_extensions~=4.10.0
pytz~=2024.1
sqlparse~=0.4.4
mysqlclient~=2.2.4
asgiref~=3.7.2
colorama~=0.4.6
PyYAML~=6.0.1
numpy~=1.26.4
MarkupSafe~=2.1.5
certifi~=2024.2.2
itypes~=1.2.0
coreapi~=2.3.3
coreschema~=0.0.4
click~=8.1.7
uritemplate~=4.1.1
urllib3~=2.2.1
inflection~=0.5.1
idna~=3.6
dj-config-url~=0.1.1